[build] Fix new -Wformat warnings
A clang change has added new warningsfor implicit
conversion of `enum class` to their underlying
type which is UB.

Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/873076
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Alex Brachet <abrachet@google.com>

+++ pw_bluetooth_sapphire/host/fidl/helpers.cc
@@ -1322,7 +1322,8 @@ ServiceDefinitionToServiceRecord(const fuchsia::bluetooth::bredr::ServiceDefinit
 
   if (definition.has_profile_descriptors()) {
     for (const auto& profile : definition.profile_descriptors()) {
-      bt_log(TRACE, "fidl", "Adding Profile %#hx v%d.%d", profile.profile_id, profile.major_version,
+      bt_log(TRACE, "fidl", "Adding Profile %#hx v%d.%d",
+             static_cast<unsigned short>(profile.profile_id), profile.major_version,
              profile.minor_version);
       rec.AddProfile(bt::UUID(static_cast<uint16_t>(profile.profile_id)), profile.major_version,
                      profile.minor_version);
